+++ src/types/index.d.ts
@@ -1,12 +1,3 @@
 import './vue'
-import { VueLogger } from './logger'
-
-export default VueLogger
-
-export {
-  LoggerOptions,
-  LoggerHook,
-  LogEvent
-} from './logger'
-
+export * from './logger'
 export * from './hooks'