+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # vue-logger-plugin changelog
 
+## [2.0.0] - 2020-09-20
+* Release for Vue 3
+    * Provide createLogger & useLogger methods
+    * Documentation updates
+
+## [2.0.0-beta.1] - 2020-09-11
+* Package updates
+
+## [2.0.0-beta.0] - 2020-08-14
+* Initial update to work with Vue 3
+* Provide createLogger & useLogger methods
+* Documentation updates
+
 ## [1.0.1] - 2020-07-22
 
 * Bump lodash from 4.17.15 to 4.17.19 (package-lock.json)
@@ -13,5 +26,8 @@
     * built-in argument stringify hooks
     * on-demand log level adjustment and enable/disable via `apply`
 
+[2.0.0]: https://github.com/dev-tavern/vue-logger-plugin/releases/tag/v2.0.0
+[2.0.0-beta.1]: https://github.com/dev-tavern/vue-logger-plugin/compare/2.0.0-beta.0...v2.0.0-beta.1
+[2.0.0-beta.0]: https://github.com/dev-tavern/vue-logger-plugin/compare/v1.0.1...2.0.0-beta.0
 [1.0.1]: https://github.com/dev-tavern/vue-logger-plugin/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/dev-tavern/vue-logger-plugin/releases/tag/v1.0.0